Expanded hash to include additional as duplicate hashes was making larger dataset impossible
The size of the chess hash was heavily limited by the parameters. The
chance of a having a duplicate opening move is nearly 100% after just a
dew data points. The lack of elos meant that we weren't getting a broad
spread of the various players elos reactions to different board states.
I also added date_time as player strategy has changes across the years.

In summary the limitations on inputs to the has limited the dataset size
significantly.