Added cuda device support to training to speed up training
Added a 'cuda_enabled' variable to config. If enabled checks to see if
the device can be used. Otherwise if the model cannot be used sends a
warning. Dataloaders now used pinned memory to speed up retrieval, and
the model, data, and criterion are now sent to device for train and
_dataset_loss.